starting of the puzzle
This is an 8-bit grayscale image that has byte-encoded data. Once decoded you will get this message:
"Problems and Answers. Humans asks problems and humans give answers. 
Some problems have the answer hidden inside them while other problems require deep research.
Sometimes you find a answer that needs to be converted into the real answer like adding x00 in a binary.
Some answers are just Problems themselves that lead to more problems. The most important thing
to answer a problem is to count every bit of the problem to understand whats going on and what is      
avilable to you to answer that problems. Put your answer here theonewhobetrayedall.onrender.com"

There are some errors in this text but who cares?